remove view from index db
