+++ packages/workspace-fusion/src/lib/utils/index.ts
@@ -6,7 +6,6 @@ export * from './sidesheet';
 export * from './statusBar';
 export * from './addCustomTab';
 export * from './createFusionWorkspace';
-export * from './viewController';
 export * from './indexedDb';
 export * from './addConfig';
-export * from './addViewController';
+export * from './viewController';
