+++ packages/workspace-fusion/src/lib/utils/viewController/index.ts
@@ -1,1 +1,1 @@
-export { addViewController } from './addViewController';
+export { addViewController, switchTabOnNavigation } from './addViewController';
